requirements.txt
numpy==1.21.5
Flask==2.0.3
Werkzeug==2.0.3
gunicorn==20.1.0
psycopg2-binary==2.9.3
requests==2.27.1
beautifulsoup4==4.10.0
pandas==1.4.2
pytz==2022.1
lxml==4.8.0